Update proc target to enforce force copy and set permissions
This change modifies the proc command in spot.yml to enforce the force copy of the mp3 file and to set the permissions of the copied mp3 file. It also adjusts proc_mp3.go and proc_mp3_test.go to reflect these changes. The verbosity flag has been removed from the spot command in proc_mp3.go and a flag for default concurrency has been added to the deploy to nodes task. Finally, the dbg flag has been removed from the proc-mp3 target in the Makefile.