Increase podcast count in RSS feed and add README
The count of podcasts included in the RSS feed has been increased from 25 to 200 to provide a wider selection of past content. Additionally, a README file has been added to the rss_generator directory to provide an overview of how the RSS feeds are built and functioning details of the rss_generator program.